build: Streamline crate names and rename the binary (#40)
Prefixes all crates with `objectstore` so that it is easier to recognize
them in the file tree and among other dependencies. Along with this, we
can now include workspace members with a wildcard.

The main binary is built by the `objectstore-server` crate and is named
`objectstore`. The dockerfile is updated to use this binary name.

The `client` crate is renamed to `objectstore_sdk`, which is a speaking
name to import in applications that use objectstore. This along with its
dependencies will be the only crate eventually published to the
registry. We will follow up with appropriate `publish = False`
annotations in the crates at a later point.

The docker builds for GHCR have also been renamed to
`getsentry/objectstore` and are now tagged with just `latest` or the
architecture (e.g. `amd64`). The `nightly` tag is removed, since we push
every commit as `latest`. This can be revised at a later point when we
enable releases via craft.