test: Add an integration test using stresstest as a library (#50)
Adds an integration test to the `objectstore-server` crate that spawns
the full server in a subprocess and then runs the stresstest against it
for 10 seconds.

To do so, the `stresstest` crate has been split into a library and
binary:

- Workload definition and the stresstest itself are in the library and
  can be called via `stresstest::run`.
- YAML configuration and a main function that reads config, configures
  workloads, and then runs the stresstest remain in the binary.

The initial stresstest integration test is not meant to mimic any real
workload and it does not run in release mode. Rather, it is used to
integration-test the stresstest with the binary.