Cleans up client stream errors, renames things (#54)
This is mostly a followup addressing some other review comments.

The primary change is that it changes the Stream Error type into
`io::Error` from `anyhow::Error`.

Otherwise, renames the `Client/Builder`, and adds a new convenience
method to pass an `AsyncRead` (such as a tokio `File` handle) as an
upload source.