feat(stresstest): Implement throughput mode (#60)
Adds a second opt-in mode for stresstests to define constant throughput
rather than constant concurrency. This allows us to test the traffic
patterns of our use case across iterations of the server with different
latency. Also, it allows us to more effectively assess the latency of
the service at specific throughput.