Always assign Id from the server (#62)
This changes the Client implementation so that object Ids are always
assigned and returned from the Server.

We would like to support multiple storage backends, and encode the used
backend into the object id. For that to work, the Id need to be
controlled by the server/service.

As a driveby change, this also simplifies the `PutBuilder` code a bit.