feat: Implement a simple backend for Google BigTable (#64)
First iteration of a Google BigTable backend that supports
 - Concurrent uploads and downloads
- Table management: Automatically creates the table, column families,
and GC policies.
- Overwriting objects: When calling `put` on an object that exists, the
old one will be deleted along with its metadata.
- TTL and TTI: Automatic expiry of objects after a time-to-live, as well
as automatic bumping via time-to-idle. The latter has to be implemented
as re-inserting, so the bump is debounced to once per day.
- Metadata storage: Encodes the metadata map, including custom metadata,
into a dedicated column.

Locally and in CI this is tested using the BigTable Emulator. The
emulator is also added to devservices for convenience. Unfortunately, we
had to bring back `protoc` since this is required by one of the
dependencies.

Future improvements:
- Chunking to store objects larger than 10 MB efficiently
- Multi-fetching of more than one object at a time
- (TBD) Splitting metadata into dedicated columns

Closes FS-107