build: Disable debug in dev again (#66)
This seems to have been removed unintentionally as a drive-by refactor.
Brings back our usual profile management that we also use in Relay and
Symbolicator.

By default, debug information is disabled in development to speed up the
build and reduce the target folder size. In a `dev-debug` profile or in
`release`, debug information is enabled.