meta: Add a virtualenv with devservices and improve README (#67)
Adds a Python virtualenv using the same Python version as Sentry. This
env includes the `devservices` tool, which we can use to run backend
services required for development and testing. At the moment, `bigtable`
is the only tool.

Also improves the README to add more instructions on environment setup
and running devservices.