build(cargo): Use workspace depependencies for shared deps (#82)
Most of our dependencies are used across crates and must match in their
versions. This PR lifts all of them up into workspace dependencies, so
that we only have a single definition.

Apart from few exceptions, features remain local to the crate that uses
the dependency. This will allow us to minimize the build once we
introduce feature flags or build smaller targets. The only features we
declare globally are the ones we expect to require everywhere.