[WIP] refactor scope (#83)
- removes the jwt stuff from client and server
- puts scope and usecase in query params instead of auth header
- the client serializes the scope in its path form when sending it in
the request. the server no longer bothers to deserialize it

existing tests pass. i tried to extend the client's `TestServer` to
allow assertions to be made on all of the requests it receives, but the
assertions had to be passed in at server creation which isn't possible
if the server can generate IDs. i will give more thought to tests
outside of this PR

i also tried to add a `Scope` newtype but i just kept needing more
boilerplate for it to behave like a string. another thing to give
thought later

i will merge as-is since it is stamped and, assuming we are aligned on
the behavior, we can avoid merge conflicts with other work (i notice for
instance that a route needs to be added for user-supplied paths)