ref(gcs): Use JSON-API instead of XML API (#85)
Adds a dedicated implementation for GCS, which switches it from the XML
to the JSON API. This API allows us to use resumable uploads, multipart
uploads, and metadata-only requests. The S3-compatible API will be left
in the code base for now, although it relies on Google-specifics in
several places and will have to be rewritten if we actually want to rely
on it.

The test server used by the GCS client libraries supports the JSON API,
which allows us to use it in CI and for local testing now. Hence, we add
it to devservices and CI for some basic unit tests.

This PR also fixes handling of TTL and TTI based on `customTime`, does
not return expired objects, properly namespaces built-in metadata keys
with typing, supports the STORAGE_EMULATOR_HOST environment variable,
and supports overriding the GCS endpoint.

Closes FS-143