ref: Revert Key/Path concepts (#86)
This reverts the recent changes to introduce a binary encoded key.

Instead, this introduces the `ObjectPath`, as the fully scoped (using
usecase and scope) path to an object. The object then has a `key` which
uniquely identifies it within the usecase/scope.

As this way, we cannot encode the backend decision in the key anymore,
instead we will query the (presumably) high-performance `high-volume`
backend first, before falling back to the long-term backend.

The thinking here is that >99% of all objects will live in the
"high-volume" backend, which is also chosen to have better performance
(lower latency) than the "long-term" backend. For that reason, misses on
the first backend are judged to have both low probability and reasonly
low overhead compared to the steep latency of that backend on its own.