ref(stresstest): Distribute workloads across usecases and orgs (#91)
Makes the stresstest more realistic by writing multiple usecases and
scopes. The name of each concurrent workload is used as usecase, which
also mimics production behavior more closely. Also, each workload can
specify a number of organizations (defaulting to 1), which is used as
scope.

Internally, we create a project-level scope, since this is what the real
usecases will also be doing. We reuse the org ID as project ID.