ref: Retry high-volume PUT failures
This adds a retry loop around PUT failures of the "high-volume" backend, which we can do because we are buffering the whole payload in memory and can thus trivially clone it.

The long-term backend however cannot be retried, as it is based on streaming uploads, which we cannot rewind.

As a driveby change, this also introduces a `name` to the generic backends, which we emit as a separate metrics tag now.