fix: Initialize Sentry before tokio (#97)
The way Sentry does its Hub/Client-propagation across threads means that
it has to be initialized on the main thread for this to work, before any
other threads are being spawned.

In other words, Sentry is notorious for being incompatible with
`#[tokio::main]`. So lets split the main function into a sync portion
initializing all the observability crates, before spawning the tokio
runtime and the main server.