fix(docker): Permissions in data volume (#103)
A docker volume is usually created with root permissions, so
`objectstore`
running as `nonroot` did not have enough permissions to create files in
the data
volume. This also happens in our devservices setup, where we mount a
volume at
`/data`.

To fix this, we first create a folder with appropriate permissions. When
mounting the empty volume for the first time, docker will inherit the
folder's
group and user.

Since there is no `mkdir` in the distroless container, we use `COPY`
from the
empty home folder in the distroless base image to create the folder.