feat(client): propagate sentry-trace header (#93)
adds an `enable_distributed_tracing` option which, if on, will populate
the `sentry-trace` header on outgoing requests to objectstore.

is it worth extra boilerplate to avoid coupling our client library with
`sentry-rust`? the "correct" way to do this is to externalize the
dependency and take an `Option<Fn() -> Option<(String, String)>>` which
is expected to return a `(header_name, trace_id)` pair. and then i guess
we give users a copy-pastable snippet or separate crate for plugging in
the specifics for Sentry.