chore: add basic debug logs and adjust span log levels (#106)
- sets our spans to `info` level so they will be sent to sentry
- cleans up our `EnvFilter` usage. we don't use the dns crate that was
referenced before, we had the wrong name for one of our targets, and
with our spans set to `info` we can disable the noise from `tower_http`.
- if we want to re-add the capability to put a level filter specifically
on printing, IMO that should be a separate env var. i think it is weird
that `RUST_LOG=debug` means `RUST_LOG=debug` for the console and
`RUST_LOG=INFO,objectstore=TRACE...` for spans.
- adds a few spans and a bunch of `debug!()`s to see a very basic
"story" when uploads pass through

the "logs" tab in sentry does not get any of the span context
surrounding the log so it is not very useful. i will look into whether
anything special has to be done for structured logging to work well in
GCP.

also will do a pass for metrics i think we want. didn't get to it today