feat: latency metric for GET/DELETE, size metric for GET requests (#110)
only implements these metrics in GCS/Bigtable, not local_fs or
s3_compatible.

regarding the size metric in GET requests: GCS will calculate the size
for us and populate it in the metadata response. Bigtable doesn't, but
we load entire Bigtable objects into memory so we compute the size and
set it in the GET handler. alternatively, we could write the size into
the metadata at PUT time so it will be deserialized automatically at GET
time, but the trait interface takes `&Metadata` and that would have to
change.

regarding deletes: if we need to speed those up, we should either make
the second backend attempt contingent on a `NOT_FOUND` from the first
attempt or run both attempts concurrently. currently you must run both
in sequence.

also, i notice the retry logic in the bigtable impl does not early-exit
for `NOT_FOUND`. we can probably decrease GET latency for objects stored
in GCS if we change that.