chore: Actually test against the real server
We keep breaking the server functionality and the stresstests.
Even though we run the stresstest in a test, it does not catch errors, as it will just report failed requests, which are swallowed in its output.

Instead, we can reuse the client tests to actually run against the *real* server, instead of a manually written mock server.