build: Exclusively build on merge_group and not main branch (#116)
Since we enabled merge queues, every merged PR was built & pushed twice.
Since the merge group commits land directly in the main branch, we can
skip building there.

This assumes that we will not have workflows bypassing the merge queue.
If we add such workflows in the future, we can revise.