ref: Remove bigtable `admin` usage, along with dependency updates
For some ops/permission related reasons, we cannot use the bigtable admin API, and we will be provisioning the needed tables/column-families through other means.

This means we can remove the related code, and also switch to a published version of that dependency. This also cleans up a couple more dependencies and runs a `cargo update` for good measure.