ref: Remove bigtable `admin` usage, along with dependency updates (#120)
We chose to change our approach to provisioning the table and column
families in BigTable. Previously, this was created by the application
and requires admin credentials. Going forward, we will instead provision
both ahead of time. The application can now assume that the table exists
and will fail if it doesn't.

This means we can remove the related admin code and also switch to a
published version of the `bigtable_rs` dependency. As drive-by, this
also cleans up a couple more dependencies and runs a `cargo update` for
good measure.

For local use, we a add a script one has to call once when the bigtable
container is created. See the README for more information.

---------

Co-authored-by: Jan Michael Auer <mail@jauer.org>