build: Adopt devenv and uv (#123)
To get up to speed with Sentry's developer tooling, we're adopting `uv`
instead of managing our python version with `pyenv`. This will
eventually become a workspace that hosts our python client as a member.

To install `uv`, we use Sentry's own `devenv`. However, `devenv` is
optional and we can also work with a globally installed version of `uv`
for now. This should make it easier to contribute. We will keep this
approach until there are more dependencies managed by `devenv`.

Where possible, we continue to use standard python tooling. Precisely,
we use `uv` over `devenv` to manage the venv.