feat(service): Add user-agent for outgoing HTTP calls (#131)
Sets the `user-agent` header to `sentry-objectstore/<version>` for
outgoing HTTP requests to GCS and S3. This should make it easier to
identify these requests in-flight and on the vendors' end.

This also centralizes building the reqwest client, which now occurs at a
single place. The backend module is slightly restructured with all
shared utilities moved into `backend::common`.

Ideally, we would also set the user agent for our bigtable client, but
there's no direct API to add a tonic interceptor on the bigtable client.
The existing `configure_inner_client` method does not expose the auth
service so we can wrap it in a new client, unfortunately. We will look
into this in a follow-up.