build: Include full debug information in release builds (#138)
Incluedes full debug information in release builds as opposed to just
line tables before. This will increase the size of the binary
significantly, but most importantly it will allow us to see inlined
functions in stack traces.

In a follow-up, we will strip debug information from the release
container and upload it to Sentry so that the container remains small.