ref(server): Create an app struct and improve inflight requests (#140)
Moves the router into an `App` struct that has a constructor and a
`serve` method that starts the `axum` server future. This also
simplifies usage in tests, as it is no longer required to import axum
directly.

Internally, `App::serve` will start all concurrent processes that the
app needs. Initially, this is just the web server and the emitter from
the in-flight requests middleware. Potentially, it is cleaner to spawn
them into separate tasks, but since we spawn externally right now I
thought it's better that this future has no side effects and can be
canceled at once.

Additionally, this improves the handling of in-flight requests. We now
use the utility from `tower-http` with a built-in drop guard for more
accurate counting. Since this requires an emitter task to run, we start
it simultaneously with the axum server and await both before finishing
the serve future.