fix(stresstest): Properly default concurrency (#145)
In a past change, the configuration for concurrency was made optional in
the stresstest to support throughput-mode. The default of `0` was still
used by cleanup with now ran with unlimited concurrency, overwhelming
the objectstore server and crashing the stresstest.

This adds the same default as the workload builder for concurrency,
which should avoid such crashes.