meta: Add cargo metadata and a client readme (#154)
Adds metadata to all crates in the workspace, most notably setting the
`publish` flag on all crates. The client is the only crate that will be
published at this point.

We also set the minimum Rust version of the client library to 1.89.0,
the second latest version at the time of writing. This additionally
configures clippy to lint at that level. We will be bumping this version
only as needed, so we do not force downstream projects to update their
toolchain prematurely.

This also adds a _README.md_ to the client and adds a usage example to
the top-level client library docstring. To simplify handling of
payloads, we introduce a `.payload()` and `.text()` helper method on the
`GetResult` type.