feat: Allow defining a per-usecase/client default expiration policy
This adds a default expiration policy to the `ClientBuilder`, which can be specified along with the usecase to set a per-usecase expiration policy.

This will make it easier to always have an expiration policy and will prevent issues as we have right now where we have never set an expiration policy for the double-write uploads we are currently doing.