feat(devservices): Initialize bigtable automatically (#159)
Changes the container startup to use our own script that automatically
creates the table and column groups. This means we no longer have to
call the script manually.

This also adds healthchecks to the bigtable and gcs containers. Since we
use a distroless container for objectstore, we cannot easily add a
healthcheck there - this will be added at a later point.