feat(server): Improve error handling during shutdown (#161)
Changes the sequence of operation during shutdown: Upon receiving the
shutdown signal, we immediately print a message on stdout. Next, we wait
for the server future to complete and log any errors. This ensures the
error is tracked in Sentry, whereas before we returned it from `main`.
Only once that is done, we flush metrics.