ref(service): Use JSON to encode metadata (#165)
Encode metadata using JSON instead of bincode. This fixes two issues:

1. Governance and evolution. We can now add and reorder metdata fields
   without risking to break past consumers.
2. Partial serialization: For cases where we need to partially serialize
   metadata, we can do so now. Bincode will fail when using `skip` on
   fields.

The implementation on LocalFs is not the most efficient one, since it
allocates a string for metdata. However, since our metadata tends to be
fairly small and we will place strict limits on the maximum number and
size of custom fields, this is OK.

This is a one-time breaking change.