build(py-client): Downgrade urllib3 requirement (#179)
`sentry` depends on an old version of `stripe` (6.7.0) and `urllib3`
(2.2.2).
Upgrading `urllib3` to 2.5.0 causes issues with a test that use `stripe`
(showing very strange behavior in my opinion), see
https://github.com/getsentry/getsentry/actions/runs/19050598423/job/54409663305
The easiest solution is to downgrade our requirement of `urllib3` to
`2.2.2` to match the version in `sentry`.

The alternative could be to venture into upgrading `urllib3` and
`stripe` in `sentry`, but given that the current version of `stripe` is
`13.1.1` (!) this would likely be quite challenging due to many breaking
changes.