ref(cli): Introduce commands and add healthcheck (#185)
Introduces commands to the CLI so we can add more functionality to the
objectstore binary than just the server:
- Docker healthchecks. Since we run in a distroless container, we need a
command internally to check for availability. In other services like
Relay and Symbolicator, this is solved with a healthcheck command.
- Future extensions: Config generation, validation, migration, ...

The webserver can now be invoked with the `run` command. Note that this
is a breaking change to our Docker container and all uses have to be
updated. This PR already updates devservices.

While restructuring the CLI bootstrap, we also further improve error
reporting of top-level command errors.