feat: Add missing PUT `key` to the Rust client, replace `anyhow` with… (#191)
… `thiserror`

We did a 360 when figuring out whether we should assign objects keys
from the server or the client. In that process, I guess I forgot to add
back the `key` to the Rust client. So here it is.

Also, this replaces our usage of `anyhow` within the `types` and
`client` crate with `thiserror`. This is a bit verbose, and I can
predict for a fact that noone will ever match on these errors or care
about which one it is.

But relay folks insist. They seem to have a "no `anyhow`" policy, as
they claim capturing backtraces on every `anyhow` that happens is
expensive. While that is true, they have also pinned a >2 year old
`anyhow` version which supposedly does not capture backtraces by
default. And relay team also does not seem aware that you can actually
disable backtraces for errors by setting `RUST_LIB_BACKTRACE=0`. But
hey, I don’t really want to argue. If they want to cargo-cult around
these decisions without challenging them every now and then, who am I to
tell them otherwise?