feat: Implement "redirect tombstones" (#186)
In order to avoid requests to the "long term" backend (aka GCS) as much
as possible, this now implements redirect tombstones, keys, or markers,
however you want to call these.

This marker is part of the "high volume" metadata for cases when the
data is actually stored in the "long term" backend.

This means that writes for which the long-term backend was chosen will
*also* write this tombstone to the high-volume backend.

On the other hand, reads and deletes which also previously were hitting
the high-volume backend first will now short-circuit when the object is
not found. Previously, reads and deletes would unconditionally cascade
to the long-term backend.