[1/n] feat(auth): add config for auth keys and associated permissions (#198)
this PR adds a new `auth` section to our config. auth has two keys:
- `enforce`: whether or not requests should return 403 when auth fails.
we can disable auth for soft rollout / test scenarios
- `keys`: configurations for keys that can be used to verify the auth
header

`keys` is a map from key id (which should correspond to the `kid` JWT
header field) to key configuration. the key configuration has the max
permissions that the key is allowed to grant and a list of versions of
the key. the key version thing allows us to rotate keys if we ever need
to, but it's not ergonomic enough to want to automate

when we want to rotate (for instance) relay's key, we:
- make sure `config.auth.keys.relay.key_versions` includes both the old
version and the new version
- deploy a new relay release which signs with the new version
- wait for the deploy to complete on all hosts
- remove the old key from our config

(apparently the tool i use for stacked PRs needs an update :( )