build(gocd): Create sentry releases during deployment (#208)
Adds a new CLI command to print the full version string that is used for
Sentry in the format of `objectstore@<version>+<revision>`. A new job
prints this release and uploads it as deployment artifact for gocd.
During deployment, we read this release name and create the release
alongside the deployment in Sentry.

This largely follows the model that Relay and Symbolicator use, with the
difference that we use regions as environment. For canaries in the
regions, we will use a `<region>-canary` suffix.