ref(server): Use "correct" methods and status codes for creation (#215)
Creating an object with or without a key usually uses different HTTP
verbs. PUT writes or overrides an object at a known key, while POST
always adds a new object with a server-defined key. The status codes
also differ, with POST returning 201 CREATED.

If we think this is useful, we could also make PUT return 201, but for
that we'd have to return from the backend whether the object previously
existed.

Internally, we now have two types:

- `ObjectPath` requires a key to be present. It is used for GET and
  DELETE.
- `OptionalObjectPath` has an optional key. It can be converted into
  `ObjectPath`, either requiring or creating a random key.

ObjectPath deserializes via OptionalObjectPath.