feat(stresstest): TTL after 1h and make cleanup configurable (#223)
Stresstests are frequently canceled or crash as we test objectstore's
limits. This leaves objects in the backends permanently. To fix this, we
now configure a default 1h TTL for all objects created by the
stresstest.

Since we can usually rely on that TTL, we no longer have to run the
cleanup. This saves significant time, as the deletions usually take
longer than the test itself.

To make configuration and usage a bit easier, the stresstest is now
converted to the command pattern.