ref(service): Make object key optional on insert (#234)
Changes the object signature and makes keys optional. This allows us to
add an optimization that skips checking for tombstones when
server-assigned keys are used.

This PR introduces a new type `ObjectContext`. It stores the usecase and
scopes of the object's ID. The `ObjectId` type now wraps a context and
has convenience constructors and getters for all its parts. This type
can be used by auth instead of the `ObjectId`, since the key part is not
needed. This will further allow us to provide **auth on batch endpoints
and collection operations**.

The `StorageService` interface now has an `insert_object` that replaces
the former `put_object`. This method takes an optional key and
internally creates a random ID if needed.