ref: move `Permission`, `Scope`, `Scopes` to `objectstore_types` and use in Rust client (#236)
Signing tokens in the client requires structured scope information which
we currently don't have. We have a structured scope type in
`objectstore_service` already so this PR moves that type to
`objectstore_types` and integrates it into the Rust client.

As a driveby, it also moves the `Permission` type in preparation for the
actual token signing impl.

---

Co-authored-by: Jan Michael Auer <mail@jauer.org>