ref(server): Create extractors for params (#238)
This creates a new module in `objectstore-server` that contains all
extractors we need for our web API.

The extractor for the auth-aware service has been moved into this
module. This gives us a cleaner separation between the HTTP layer and
the auth logic, which would even allow us to move that wrapper into a
different crate. Additionally, there are minor changes to it:

- The service no longer has an explicit `enforce` flag. This is inferred
  from `Option` and documented on its new constructor.
- We now support case-insensitive parsing of the bearer prefix.

There are also two new extractors for `ObjectId` and `ObjectContext`.
Endpoints no longer have to use the Params types directly and also do
not have to deal with configuring the Sentry scope.

This change is a preparation for introducing killswitches.