feat: Implement static killswitches (#239)
This implements a first version of killswitches. Switches are configured
through objectstore config, so in order to create a killswitch one has
to apply new config and restart the service.

Note that all scope values must be strings. Examples for configuration
look like this:

```yaml
killswitches:
  # Drop all requests for a specific usecase, regardless of scopes
  - usecase: broken_usecase

  # Drop requests from a specific org, all projects, all usecases
  - scopes:
      org: "42"

  # Drop requests from a specific project, all usecases
  - scopes:
      org: "42" # can also be omitted
      project: "4711"

  # Combination of usecase and org
  - usecase: attachments
    scopes:
      org: "42"
```

At a later stage, we can replace the static configuration with a runtime
system.