ref(clients): Change defaults for timeouts (#246)
Our read timeouts have been causing issues on the server side since they
cause the connection to drop, which can leave operations in a dangling
state. This has to be addressed separately, but also uncovered that a
default read timeout is likely not desired by use cases.

This changes timeout functionality in both clients equally:

1.  There is a default connect timeout of 100ms
2.  There is no default read timeout anymore
3.  The `timeout` / `timeout_ms` setting only affects read timeouts
4.  To set connection timeouts, one has to use the lower-level client
    config.

This is a BREAKING CHANGE