ref(server): Reorganize the HTTP web module (#247)
Splits the HTTP module internally into smaller submodules that should
make it easier to find and place functionality. The web listener, app,
and middlewarers are now in separate modules.

Additionally, `ApiResult` and its error has moved into an
`endpoints::common` module. We can later replace this with an error type
with individual variants.